pyproject.toml
[project]
name = "anyway"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "boto3>=1.42.77",
    "fastapi>=0.135.2",
    "google-genai>=1.68.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "openai>=2.30.0",
    "piexif>=1.1.3",
    "pillow>=12.1.1",
    "pillow-heif>=1.3.0",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
